Fixes for the EPS versions of some Oxygen icons used in the relax manual.
This is the actions.document-preview-archive and actions.office-chart-pie Oxygen icons used for the
user function icons.  The files were not created correctly in the Gimp.  The export to EPS requires
the width and height to be both set to 6 mm, and the X and Y offsets to zero.  This allows the icon
bounding boxes and sizes to match the other EPS icons.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27551 b7916896-f9f9-0310-9fe5-b3996d8957d5