Big bug fix for the N-state model num_data_points() function.
This is from the specific_analyses.n_state_model.data module.  This code was very much out of date.
It was expecting an ancient behaviour where the spin container 'pcs' variable and interatomic data
container 'rdc' where lists of floats.  However these were converted many years ago to dictionaries
with keys set to the alignment IDs.  The result was that no RDCs nor PCSs were counted as a base
data point, so the function would in most cases return a value of zero.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27561 b7916896-f9f9-0310-9fe5-b3996d8957d5