Implementation of the statistics.aic user function.
This is very similar to the statistics.model user function - the code was copied and only slightly
modified.  The new user function will calculate the current chi-squared value per model, obtain the
model statistics, calculate the AIC value per model, and store the AIC value, chi-squared value and
number of parameters in the appropriate location for the model in the relax data store.



git-svn-id: svn://svn.gna.org/svn/relax/trunk@27569 b7916896-f9f9-0310-9fe5-b3996d8957d5